Refactors get file functions in meta compress
