+++ cg/meta/compress/__init__.py
@@ -1,0 +1,1 @@
+from .compress import CompressAPI
